+++ axelor-human-resource/src/main/java/com/axelor/apps/hr/db/repo/TeamTaskHRRepository.java
@@ -33,17 +33,14 @@ public class TeamTaskHRRepository extends TeamTaskProjectRepository {
     super.save(teamTask);
 
     teamTask.setTotalPlannedHrs(projectPlanningTimeService.getTaskPlannedHrs(teamTask));
-    teamTask.setTotalRealHrs(projectPlanningTimeService.getTaskRealHrs(teamTask));
 
     Project project = teamTask.getProject();
     project.setTotalPlannedHrs(projectPlanningTimeService.getProjectPlannedHrs(project));
-    project.setTotalRealHrs(projectPlanningTimeService.getProjectRealHrs(project));
 
     Project parentProject = project.getParentProject();
     if (parentProject != null) {
       parentProject.setTotalPlannedHrs(
           projectPlanningTimeService.getProjectPlannedHrs(parentProject));
-      parentProject.setTotalRealHrs(projectPlanningTimeService.getProjectRealHrs(parentProject));
     }
 
     return teamTask;
@@ -56,7 +53,6 @@ public class TeamTaskHRRepository extends TeamTaskProjectRepository {
     super.remove(teamTask);
 
     project.setTotalPlannedHrs(projectPlanningTimeService.getProjectPlannedHrs(project));
-    project.setTotalRealHrs(projectPlanningTimeService.getProjectRealHrs(project));
   }
 
   @Override