+++ axelor-human-resource/src/main/java/com/axelor/apps/hr/service/project/ProjectPlanningTimeService.java
@@ -28,11 +28,7 @@ public interface ProjectPlanningTimeService {
 
   public BigDecimal getTaskPlannedHrs(TeamTask teamTask);
 
-  public BigDecimal getTaskRealHrs(TeamTask teamTask);
-
   public BigDecimal getProjectPlannedHrs(Project project);
-
-  public BigDecimal getProjectRealHrs(Project project);
 
   public void addMultipleProjectPlanningTime(Map<String, Object> dataMap) throws AxelorException;