+++ CHANGELOG.md
@@ -141,6 +141,7 @@
 - SUPPLYCHAIN : Auto-complete sale order when it has been all invoiced
 - Stock: Add stock history view for a given product, company and stock location.
 - Production : Provide menu for Machine Planning.
+- Teamtask : Removing real hours management from project planning time
 
 ## Improvements
 - Contract: added button to manually close contract if termination date was set in the future